#10. Spring Boot FlyWay: миграции БД, профиль пользователя (https://www.youtube.com/watch?v=ArM7nCys4hY&list=PLU2ftbIeotGoGSEUf54LQH-DgiQPF2XRO&index=10)

+++ src/main/kotlin/com/example/sweater/domain/Message.kt
@@ -7,7 +7,7 @@ import javax.persistence.*
 data class Message(
         @Id
         @GeneratedValue(strategy = GenerationType.AUTO)
-        val id: Int,
+        val id: Long,
         val text: String,
         val tag: String,
         val filename: String?,
@@ -17,6 +17,6 @@ data class Message(
         val author: User
 )
 
-interface MessageRepo : CrudRepository<Message, Int> {
+interface MessageRepo : CrudRepository<Message, Long> {
         fun findByTag(tag: String): List<Message>
 }
